Fix lefthook capturing stdin from terminal even if when non-interactive and not using stdin
If I'm reading the code, this `args.in` should be a nullReader when
non-interactive and not using stdin. For some reason though, it ends up
capturing stdin. That is, any typing you do while lefthook is running
goes to lefthook, instead of the shell invoking it.

Fixes https://github.com/evilmartians/lefthook/issues/588