Update uniseg
I don't understand how it could possibly be messing with stdin, but
uniseg 0.2.0, which is a dependency of lipgloss, which we use for nicely
colored logs... is causing stdin to be captured.

Originally, I thought it was lipgloss, but there weren't any a huge
diff, and nothing immediately jumped out. I did notice uniseg is a
dependency, and was also updated. So I upgraded it on a whim, and it no
longer swallos stdin :-|