fix(git): make short status parser more robust (#1236)
Switch to the NUL separated `-z` porcelain format recommended for
machine parsing to make it easier to avoid hitting issues with filenames
containing spaces or ones the non-z format would quote, as well as
pathological cases like ones containing ` -> `, etc.

A downside is that parsing this format is stateful and we need to "know"
the cases for which two filenames are output, currently copy (C) and
rename (R).

Ref https://git-scm.com/docs/git-status#_short_format